buffwork 00
시전자를 기준으로 연동은 했으며, 스킬 구현 및 버프적용을 하면 된다.

+++ ActionMake00/Assets/Scripts/Character/Enemy/Node/EnemyNode/EnemyTest.cs
@@ -53,7 +53,7 @@ public class EnemyTest : FollwingPlayerEnemyBT
         }
     }
 
-    public override void TakeDamage(int amount, int hitLevel = -1)
+    public override void TakeDamage(float amount, int hitLevel = -1)
     {
         base.TakeDamage(amount, hitLevel);
         if (isSuperArmor == false || hitLevel != -1)
